Updated install script, added autoftpd files and tweaks
Added a configuration `LED_DISABLED` to `autoftp.defaults.conf` to
allow to disable operation of LEDs (as this project is configured
against the LEDs available on the gl-AR300M, and not necessarily on
any other router.

Added all of the required files, except for the `login.cfg` file,
which should be placed at `/etc/autoftp/login.cfg`.

Completed installation script.